+++ src/main/java/com/GGomDDakPing/QnLove/QnLove/posts/Post.java
@@ -1,0 +1,49 @@
+package com.GGomDDakPing.QnLove.QnLove.posts;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import org.hibernate.annotations.CreationTimestamp;
+import org.springframework.data.annotation.LastModifiedDate;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Post {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne
+  @JoinColumn(name = "memberId", nullable = false)
+  private Member member;
+
+  @Column(nullable = false)
+  private String title;
+
+  @Column(nullable = false)
+  private String content;
+
+  @Column(nullable = false)
+  private String answer;
+
+  @CreationTimestamp
+  private LocalDateTime createdAt;
+
+  @LastModifiedDate
+  private LocalDateTime updatedAt;
+
+  @Column(nullable = false)
+  private boolean isDeleted = false;
+
+  @OneToMany(mappedBy = "post", cascade = CascadeType.ALL, orphanRemoval = true)
+  private List<Quiz> quiz;
+
+}
